+++ Docs/paper.bib
@@ -39,7 +39,8 @@
 	pages = {99126U},
 	publisher = {Proc SPIE},
 	abstract = {<p>The Daniel K Inouye Solar Telescope (DKIST) polarimetric instrumentation requires very high accuracy calibration of a complex coude path with an off-axis f/2 primary mirror, time dependent optical configurations and substantial field of view to accomplish the critical science observations. Accurate polarization predictions across a diversity of optical configurations, tracking scenarios, slit geometries and vendor coating formulations are critical to the telescope and instrument development efforts for both the design and fabrication process as well as during calibrations and post-processing mitigation efforts. Recent successesin daytime sky polarization calibrations and Zemax system polarization modeling of the 4m AEOS telescope and HiVIS spectropolarimeter show feasibility of making reasonably accurate system performance predictions. AEOS and HiVIS are a DKIST analog with a many-fold coude optical feed creating 100\% polarization cross-talk with altitude, azimuth and wavelength. Predictions of polarimetric performance for the DKIST first-light instrumentation suite have been created under a range of configurations. These new modeling tools and predictions have substantial impact for the design and calibration process in the presence of manufacturing issues, science use-case based performance predictions and ultimate system limitations in the complex DKIST optical train.</p>},
-	author = {Harrington, David and Sueoka, Stacey R}
+	author = {Harrington, David and Sueoka, Stacey R},
+    doi = {10.1117/12.2233176}
 }
 
 
