### 2020-08-26 262157f
Edited paper
