+++ Docs/paper.md
@@ -50,6 +50,6 @@ Results of the code have been verified against previous works (@Pruthvi2018) and
 
 # Acknowledgements
 
-This work has been carried out as a part of ongoing project `Jets in the solar atmosphere`, funded by the Deutsche Forschungsgemeinschaft (DFG, German Research Foundation) - Projektnummer 407727365. I thank DVS Phanindra (IIA, Bengaluru) and V. Sreekanth Reddy (CHESS, Hyderabad) for the discussion, and Mathias Waidele (KIS, Freiburg) for his feedback.
+This work has been carried out as a part of ongoing project _Jets in the solar atmosphere_, funded by the Deutsche Forschungsgemeinschaft (DFG, German Research Foundation) - Projektnummer 407727365. I thank DVS Phanindra (IIA, Bengaluru) and V. Sreekanth Reddy (CHESS, Hyderabad) for the discussion, and Mathias Waidele (KIS, Freiburg) for his feedback.
 
 # References
