### 2020-08-26 1771b4f
Edited paper
