### 2020-09-01 #2602 @pibion
@kthyng I have a lot of experience with simulating beam polarization in accelerator systems, which maps weirdly well onto modeling optical systems.  But I've never applied that knowledge outside of beam modeling.

I'm comfortable being the editor, although I'd go for three reviewers rather than two.

How does that sound?  Totally understand if you'd rather have someone with telescope-specific experience here.
