### 2020-09-12 #2602 @pibion
Excellent @caldarolamartin , @aquilesC , I've added you as reviewers!  We need one more reviewer for this paper and then I'll begin the review.  Once that happens, a new "review" issue will be opened and you'll both get a checklist in the issue.  The review process will involve everyone going through their review checklists.
