### 2020-09-15 #2602 @pibion
:wave: @garrettj403, @cavestruz, would any of you be willing to review this submission for JOSS? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/review_criteria.html
