### 2020-09-24 #2693 @whedon
[REVIEW]: PyAstroPol : A Python package for the instrumental polarization analysis of the astronomical optics.
**Submitting author:** @hemanthpruthvi (<a href="http://orcid.org/0000-0002-4892-6561">Hemanth Pruthvi</a>)
**Repository:** <a href="https://github.com/hemanthpruthvi/PyAstroPol" target ="_blank">https://github.com/hemanthpruthvi/PyAstroPol</a>
**Version:** v0.1.0
