### 2020-09-25 #2693 @caldarolamartin
Hi. I cannot modify the check list and the link to accept is expired. @pibion could you please give me a hand to solve this? Thanks
