### 2020-09-28 #2693 @pibion
@caldarolamartin, @mwcraig are you signed in to github?  You should only need to (1) subscribe to the issue and (2) be signed in to edit the checkboxes.

@arfon is that correct?
