### 2020-10-06 d2a6a75
Added contributing description
