### 2020-10-07 d513003
Added Zemax comparison, documentation
