### 2020-10-07 cd3f756
Tweaked paper
