### 2020-10-12 #2693 @pibion
@hemanthpruthvi, I think the idea of an environment file is a good one since the dependencies are somewhat sensitive.  @caldarolamartin , I've always used conda environment files for this.  Do you have any other recommendations?
