### 2020-10-13 #2693 @aquilesC
@pibion and @hemanthpruthvi , I am in the middle of the review. I strongly suggest, as @caldarolamartin points out, to create requirement files and a setup file. This will make life much easier for other users. Sorry for the self-promotion, I've written a guide on setup files [here](https://www.pythonforthelab.com/blog/how-create-setup-file-your-project/), and a very short introduction to virtual environments [here](https://www.pythonforthelab.com/blog/virtual-environment-is-a-must-have-tool/) that can be used as a starting point.

Even though conda environments are very handy, they are particularly well suited for isolating complete work environments. I am not completely convinced conda environment files are meant to be distributed to a wider audience unless there are stringent needs (and this is not the case for PyAstroPol that relies on widely available libraries). I believe that a conda environment.yml file may create conflicts (conda environments are not particularly flexible). I would favor a requirements.txt file that makes explicit required (minimum) versions.
