### 2020-10-13 #2693 @hemanthpruthvi
@aquilesC  Thank you very much for the thorough review and the valuable comments, I am working on addressing them. Meanwhile, I want to quickly address the issue regarding the **Material Licenses**. If I understand correctly, there are two issues.
1. Can the Material files be re-distributed, according to their existing licensing?
2. If so, can they be re-distributed under MIT license?

For issue 1 :
Albeit the information listed in the Material files originated in various publications, the raw files themselves are directly downloaded from [https://refractiveindex.info/](https://refractiveindex.info/). The sources of the data are also listed there. The website's database is public domain via CC0 1.0 license i.e., whatever data hosted there is free to be used/distributed without seeking any consent.

![image](https://user-images.githubusercontent.com/69858702/95852393-51d04000-0d71-11eb-8273-932a94a770d2.png)

For issue 2 :
If re-distributing under MIT license is an issue, I propose either of two ways for resolution, whichever is plausible.
1. Having two licenses : MIT -- for the code, and CC0 1.0 -- for the database, or
2. Exclude the database from the  scope of the existing license, and simply state --
"XXX data is downloaded from YYY, hosted under ZZZ license"
