+++ Materials/0_README.md
@@ -1,0 +1,20 @@
+## References
+Although the materials are downloaded from [https://refractiveindex.info/](https://refractiveindex.info/), the original source of the data is very much different. They are listed here for the sake of reference and reproducability.
+
+**Ag** and **Al** : 
+McPeak et.al., 2015 ([DOI](https://doi.org/10.1021/ph5004237)).
+
+**FusedSilica** : 
+Malitson, 1965 ([DOI](https://doi.org/10.1364/JOSA.55.001205)).
+
+**SodaLime** : 
+Rubin, 1985 ([DOI](https://doi.org/10.1016/0165-1633(85)90052-8)).
+
+**Zerodur&reg;*** : 
+SCHOTT, TIE-43 ([Source](https://refractiveindex.info/download/data/2007/schott_tie-43_optical_properties_of_zerodur_november_2007_us.pdf)).
+
+**N-XXXXX*** glasses : 
+SCHOTT catalog ([Source](https://www.schott.com/d/advanced_optics/ac85c64c-60a0-4113-a9df-23ee1be20428/1.17/schott-optical-glass-collection-datasheets-english-may-2019.pdf)).
+
+
+\* The refractive index values are computed by using Sellmeier dispersion model with coefficents stated by the manufacturer.
