### 2020-10-14 #2693 @hemanthpruthvi
@aquilesC  I understand and I agree. Accordingly, I have added a Readme file in the materials folder that refers to original sources.
