### 2020-10-14 b42a426
Modified paper
