+++ Docs/paper.md
@@ -23,33 +23,35 @@ bibliography: paper.bib
 
 # Statement of Need
 
-Polarization analysis is one of the key aspects of the optical system analysis in astronomy. Majority of the optical surfaces in astronomical instruments induce a change in state of polarization of incoming light, which could lead to inaccurate measurements of state of polarization or _polarimetry_. Hence, polarimetric instruments inevitably require the information of polarization properties of the optical system – _instrumental polarization_. It can be determined by means of experiments, analysis or a combination of both. As the telescopes become larger and instruments become more complex, instrumental polarization analysis has become all the more crucial (e.g.,  @TMT_2015 ; @DKIST_2016).  
+Polarization analysis is one of the key aspects of the optical system analysis in astronomy. The majority of the optical surfaces in the astronomical instruments induce a change in the state of polarization of the incoming light. This could lead to inaccurate measurements of the state of polarization or _polarimetry_. Hence, polarimetric instruments inevitably require the information of polarization properties of the optical system or the _instrumental polarization_. It can be determined by means of experiments, analysis or a combination of both. As the telescopes become larger and instruments become more complex, the instrumental polarization analysis has become all the more crucial (e.g.,  @TMT_2015 ; @DKIST_2016).  
 
 # Summary
 
 The Python package `PyAstroPol` provides a means to analyze the polarization properties of a given optical system with relative ease and very limited dependencies. The one simple goal is to calculate the Mueller matrix (e.g., @Gil_2016_MullMatAppr) of the optical system.  
-It can be easily distributed along with polarization calibration routines. Albeit there are several limitations, the available features should cover a large set of polarization analysis related applications. The limitations and salient features are listed below.
+It can be easily distributed along with the polarization calibration routines. Albeit there are several limitations, the available features should cover a large set of the polarization analysis related applications. The limitations and salient features are listed below.
 
 Important Limitations :   
 - All the analysis uses strictly ray treatment. Hence, all the limitations of the rays optics shall be applicable.   
-- Only circular optics (apertures) can be devised at the moment. Elliptical and rectangular optics are not possible.   
+- Only circular optics (apertures) can be devised at the moment.   
 - Incidentally, birefringent components are not included yet. The justification for this choice is that the behavior of the birefringent components is fairly straightforward as they strongly polarize the light.   
-- Visualization features are limited as the goal of this package is to keep things as simple as possible. Hence, widely used `matplotlib` package is used for the graphics. All the optics, rays, directions of the coordinate system can be displayed but pan and zoom features are severely limited.   
-- The package can be described as collection of the important objects used in optics. It is neither interactive nor design-oriented. It means that user must know the optical system beforehand, and every time a parameter changes user must manually rerun the analysis. However, it should be possible to automate this process with a little effort.   
+- Limited visualization features.
+- It is neither a design software nor interactive. That is, the user must know the optical system that is to be analyzed and the system must be updated everytime a component is changed.
 
 Salient features :   
-- Calculation of Mueller matrix of the given optical system, which is a simple and only end goal. All the rest are by-products of such analysis. However, a major caveat here is that, to compute Mueller matrix the electric fields of all the rays will be added coherently after propagating through the system. It implies that the system is imaging type, which is commonplace in astronomy.   
-- Astronomical source can be placed directly, using relevant coordinates namely declination, hour angle and latitude of the site.   
[…]
