+++ Docs/paper.md
@@ -23,32 +23,32 @@ bibliography: paper.bib
 
 # Statement of Need
 
-Polarization analysis is one of the key aspects of the optical system analysis in astronomy. The majority of the optical surfaces in the astronomical instruments induce a change in the state of polarization of the incoming light. This could lead to inaccurate measurements of the state of polarization or _polarimetry_. Hence, polarimetric instruments inevitably require the information of polarization properties of the optical system or the _instrumental polarization_. It can be determined by means of experiments, analysis or a combination of both. As the telescopes become larger and instruments become more complex, the instrumental polarization analysis has become all the more crucial (e.g.,  @TMT_2015 ; @DKIST_2016).  
+Polarization analysis is one of the key aspects of the optical system analysis in astronomy. The majority of the optical surfaces in the astronomical instruments induce a change in the state of polarization of the incoming light, which could lead to inaccurate measurements of the state of polarization or _polarimetry_. Hence, polarimetric instruments inevitably require the information of polarization properties of the optical system or the _instrumental polarization_. It can be determined through experiments, analysis or a combination of both. As the telescopes become larger and instruments become more complex, the instrumental polarization analysis has become all the more crucial (e.g.,  @TMT_2015 ; @DKIST_2016).  
 
 # Summary
 
-The Python package `PyAstroPol` provides a means to analyze the polarization properties of a given optical system with relative ease and very limited dependencies. The one simple goal is to calculate the Mueller matrix (e.g., @Gil_2016_MullMatAppr) of the optical system.  
+The Python package `PyAstroPol` provides a means to analyze the polarization properties of a given optical system with relative ease and minimal dependencies. The one simple goal is to calculate the Mueller matrix (e.g., @Gil_2016_MullMatAppr) of the optical system.  
 It can be easily distributed along with the polarization calibration routines. Albeit there are several limitations, the available features should cover a large set of the polarization analysis related applications. The limitations and salient features are listed below.
 
 Important Limitations :   
-- All the analysis uses strictly ray treatment. Hence, all the limitations of the rays optics shall be applicable.   
+- All the analysis uses strictly the ray treatment. Hence, all the limitations of the rays optics shall be applicable.   
 - Only circular optics (apertures) can be devised at the moment.   
-- Incidentally, birefringent components are not included yet. The justification for this choice is that the behavior of the birefringent components is fairly straightforward as they strongly polarize the light.   
+- Incidentally, birefringent components are not included yet. The justification for this choice is that the behaviour of the birefringent components is fairly straightforward as they strongly polarize the light.   
 - Limited visualization features.
-- It is neither a design software nor interactive. That is, the user must know the optical system that is to be analyzed and the system must be updated everytime a component is changed.
+- It is neither a design software nor interactive. That is, the user must know the optical system that is to be analyzed and the system must be updated every time a component is changed.
 
 Salient features :   
-- Calculation of the Mueller matrix of a given optical system, which is a simple and only end goal. All the rest are by-products of such analysis. The system Mueller matrix is computed by means of coherently adding the electric field vectors after propagation. It implies that the system is imaging type, which is commonplace in astronomy.   
-- Astronomical sources can be directly placed using relevant coordinates namely, declination, hour angle and latitude of the telescope site.   
+- Calculation of the Mueller matrix of a given optical system is a simple and only end goal. All the rest are by-products of such analysis. The system Mueller matrix is computed by coherently adding the electric field vectors after propagation. It implies that the system is imaging type, which is commonplace in astronomy.   
+- Astronomical sources can be directly placed using relevant coordinates, namely, declination, hour angle and latitude of the telescope site.   
 - Off-axis components are facilitated, as they have a significant effect on polarization.   
[…]
