+++ Docs/paper.md
@@ -23,35 +23,37 @@ bibliography: paper.bib
 
 # Statement of Need
 
-Polarization analysis is one of the key aspects of the optical system analysis in astronomy. The majority of the optical surfaces in the astronomical instruments induce a change in the state of polarization of the incoming light, which could lead to inaccurate measurements of the state of polarization or _polarimetry_. Hence, polarimetric instruments inevitably require the information of polarization properties of the optical system or the _instrumental polarization_. It can be determined through experiments, analysis or a combination of both. As the telescopes become larger and instruments become more complex, the instrumental polarization analysis has become all the more crucial (e.g.,  @TMT_2015 ; @DKIST_2016).  
+The instrumental polarization analysis is one of the key aspects of the optical system analysis of the astronomical telescopes and instruments. The majority of the optical surfaces in the astronomical instruments induce a change in the state of polarization of the incoming light, which could lead to inaccurate measurements of the state of polarization or _polarimetry_. Hence, polarimetric instruments inevitably require the information of polarization properties of the optical system or the _instrumental polarization_. It can be determined through experiments, analysis or a combination of both. As the telescopes become larger and instruments become more complex, the instrumental polarization analysis has become all the more crucial (e.g.,  @TMT_2015 ; @DKIST_2016).  
 
 # Summary
 
-The Python package `PyAstroPol` provides a means to analyze the polarization properties of a given optical system with relative ease and minimal dependencies. The one simple goal is to calculate the Mueller matrix (e.g., @Gil_2016_MullMatAppr) of the optical system.  
-It can be easily distributed along with the polarization calibration routines. Albeit there are several limitations, the available features should cover a large set of the polarization analysis related applications. The limitations and salient features are listed below.
+The Python package `PyAstroPol` provides a means to analyze the polarization properties of a given optical system with relative ease and minimal dependencies. The one simple goal is to calculate the Mueller matrix (e.g., @Gil_2016) of the optical system.  
+It can be easily distributed along with the polarization calibration routines. 
+
+In the polarization analysis of the astronomical telescopes, various approaches have been adopted considering the complexity of the system. A significant part of the complexity is due to the fact that the instrumental polarization is often time-dependent. For the solar telescopes with Coelostats, Mueller matrices were analytically derived as a function of time (e.g., @KTT_1985, @VTT_2005). For Thirty Meter Telescope (TMT), a combination of analytical and numerical methods is used (@TMT_2015). However, for Daniel K. Inouye Solar Telescope (DKIST), Zemax – a commercial software is used, along with the team's in-house tools (@DKIST_2016).
+
+`PyAstroPol` aims to provide better parts of the aforementioned approaches: open-source tools developed using scientific programming language, with a range of applications in modelling the astronomical optics. There is no open-source software which has polarization ray propagation features as per my knowledge. Hence, Zemax OpticStudio&reg; modelling is used for the comparison. The salient features and important limitations are listed below.
+
+Salient features :   
+- Calculation of the Mueller matrix of a given optical system is a simple and only end goal. Rest are the by-products of such analysis. The system Mueller matrix is computed by coherently adding the electric field vectors after propagation. It implies that the system is imaging type, which is commonplace in astronomy.   
+- Astronomical sources can be directly placed in the model using relevant coordinates, namely, declination, hour angle and latitude of the telescope site.   
+- Off-axis components are facilitated, as they have a significant effect on polarization.   
+- Effect of multi-layered coatings, such as oxide layers and protective coatings, on the state of polarization, is included. These are also significant in the polarization analysis of the astronomical optics (e.g., @VanHarten_2009).   
[…]
