+++ Docs/paper.bib
@@ -44,7 +44,7 @@
 }
 
 
-@BOOK{Gil_2016_MullMatAppr,
+@BOOK{Gil_2016,
  author = {{Gil}, J.~J. and {Ossikovski}, R.},
  title = "{Polarized Light and the Mueller Matrix Approach}",
  keywords = {MUELLER MATRIX,POLARIZATION},
@@ -55,7 +55,7 @@
 }
 
 
-@article{Pruthvi2018,
+@article{Pruthvi_2018,
     author = {Pruthvi, Hemanth and Krishnappa, Nagaraju and Belur, Ravindra and Kadagattor, Ramesh Bheemarao},
     doi = {10.1117/1.JATIS.4.4.048002},
     file = {:C$\backslash$:/Users/pruthvi/Documents/Mendeley Desktop/Pruthvi et al. - 2018 - Solar spectropolarimetry of Ca II 8542 Å line polarimeter development, calibration, and preliminary observations.pdf:pdf},
@@ -70,4 +70,34 @@
     url = {https://www.spiedigitallibrary.org/journals/Journal-of-Astronomical-Telescopes-Instruments-and-Systems/volume-4/issue-04/048002/Solar-spectropolarimetry-of-Ca-II-8542Å-line--polarimeter-development/10.1117/1.JATIS.4.4.048002.full},
     volume = {4},
     year = {2018}
+}
+
+
+@ARTICLE{KTT_1985,
+ author = {{Balasubramaniam}, K.~S. and {Venkatakrishnan}, P. and {Bhattacharyya}, J.~C.
+ },
+ title = "{Measurement of vector magnetic fields. I - Theoretical approach to the instrumental polarisation of the Kodaikanal Solar Tower}",
+ journal = {\solphys},
+ keywords = {Magnetic Measurement, Optical Polarization, Polarimetry, Solar Instruments, Telescopes, Zeeman Effect, Force Vector Recorders, Line Spectra, Mirrors, Optical Reflection, Polarization Characteristics, Solar Spectrometers, Stokes Theorem (Vector Calculus)},
+ year = 1985,
+ month = sep,
+ volume = 99,
+ pages = {333-348},
+ doi = {10.1007/BF00157317},
+ adsurl = {http://adsabs.harvard.edu/abs/1985SoPh...99..333B},
+ adsnote = {Provided by the SAO/NASA Astrophysics Data System}
+}
+
+@article{VTT_2005,
+abstract = {It is essential to properly calibrate the polarimetric properties of telescopes, if one wants to take advantage of the capabilities of high preci- sion spectro-polarimeters. We have constructed a model for the German Vacuum Tower Telescope (VTT) that describes its time-dependent polarization properties. Since the coelostat of the telescope changes the polarization state of the light by introducing cross talk among different polarization states, such amodel is necessary to correct themeasurements, in order to retrieve the true polarization as emitted fromthe Sun. The telescope model is quantified by a time-dependent Mueller matrix that depends on the geometry of the light beam through the telescope, and on material properties: the refractive indices of the coelostat mirrors, and the birefringence of the entrance window to the vacuum tube. These material properties were determined experimentally in-situ by feeding the telescope with known states of polarization (including unpolarized light) and by measuring its response, and from measurements of an aluminum-coated sample in the laboratory. Accuracy can in our case be determined only for the combination of telescope and spectro-polarimeter used; for the instrument POLIS at the VTT, we estimate an accuracy of ±4–5 × 10−3 for the cross talk correction coefficients.},
+author = {Beck, C. and Schlichenmaier, Rolf and Collados, Manolo and {Bellot Rubio}, L. and Kentischer, T.},
[…]
