### 2020-10-16 418a91c
Modified documentations
