+++ Docs/PyAstroPol.rst
@@ -1,70 +1,0 @@
-PyAstroPol package
-==================
-
-Submodules
-----------
-
-PyAstroPol.Coating module
--------------------------
-
-.. automodule:: PyAstroPol.Coating
-    :members:
-    :undoc-members:
-    :show-inheritance:
-
-PyAstroPol.Elements module
---------------------------
-
-.. automodule:: PyAstroPol.Elements
-    :members:
-    :undoc-members:
-    :show-inheritance:
-
-PyAstroPol.Functions module
----------------------------
-
-.. automodule:: PyAstroPol.Functions
-    :members:
-    :undoc-members:
-    :show-inheritance:
-
-PyAstroPol.Material module
---------------------------
-
-.. automodule:: PyAstroPol.Material
-    :members:
-    :undoc-members:
-    :show-inheritance:
-
-PyAstroPol.Source module
-------------------------
-
-.. automodule:: PyAstroPol.Source
-    :members:
-    :undoc-members:
-    :show-inheritance:
-
-PyAstroPol.Surface module
--------------------------
-
-.. automodule:: PyAstroPol.Surface
-    :members:
-    :undoc-members:
-    :show-inheritance:
-
-PyAstroPol.System module
-------------------------
-
-.. automodule:: PyAstroPol.System
-    :members:
-    :undoc-members:
-    :show-inheritance:
-
-
-Module contents
----------------
-
-.. automodule:: PyAstroPol
-    :members:
-    :undoc-members:
-    :show-inheritance:
