+++ Docs/index.rst
@@ -1,20 +1,0 @@
-.. PyAstroPol documentation master file, created by
-   sphinx-quickstart on Mon Aug 17 12:34:06 2020.
-   You can adapt this file completely to your liking, but it should at least
-   contain the root `toctree` directive.
-
-Welcome to PyAstroPol's documentation!
-======================================
-
-.. toctree::
-   :maxdepth: 2
-   :caption: Contents:
-
-
-
-Indices and tables
-==================
-
-* :ref:`genindex`
-* :ref:`modindex`
-* :ref:`search`
