+++ requirements.txt.txt
@@ -1,0 +1,2 @@
+numpy>=1.18.5
+matplotlib>=3.2.2
