+++ Docs/docs/modules.rst
@@ -1,0 +1,7 @@
+PyAstroPol
+==========
+
+.. toctree::
+   :maxdepth: 4
+
+   PyAstroPol

[…]
