### 2020-10-17 1af7954
Modified documentation
