+++ README.md
@@ -3,9 +3,9 @@ Instrumental Polarization Analysis of Astronomical Optics
 
 ## Overview
 The package has one simple goal : compute 4x4 Mueller matrix for the given optical system, and it is developed keeping astronomical optics in view.
-It uses geomatric optics approach i.e., all the analysis uses strictly ray treatment. Users should keep in mind that this is NOT for optical design i.e., it is presumed that the user already knows the optical system that is to be analyzed.
+It uses geometric optics approach i.e., all the analysis uses strictly ray treatment. Users should keep in mind that this is NOT for optical design i.e., it is presumed that the user already knows the optical system that is to be analyzed.
 
-The package imports following external libraries, all of which are ubiquitous. They accompany any decent scientific `Python` distribution hence this package should function with virtually no dependency issues __however__ it should be noted that it is developed with `Python3.6`.
+The package imports following external libraries, all of which are ubiquitous. They accompany any decent scientific `Python` distribution hence this package should function with virtually no dependency issues **however** it should be noted that it is developed with `Python3.6`.
 ```python
 numpy
 matplotlib
@@ -19,7 +19,7 @@ Follow these steps to start using the package.
 
 1. Go to the user directory where the package is to be installed.  
 `cd <User directory>`   
-Download the package using    
+Download the package from the Github.   
 `git clone https://github.com/hemanthpruthvi/PyAstroPol.git`  
 Rename the top directory from `PyAstroPol.git` to `PyAstroPol`
 
@@ -32,14 +32,10 @@ In Linux systems, this can be done with the command line
 `export PYTHONPATH=<User directory>/PyAstroPol`
 
 4. Import this package to your `Python` script using   
-`import PyAstroPol`  
-or  
 `import PyAstroPol as pap`  
-or   
-`from PyAstroPol import *`
 ### Examples
 
-[PyAstroPol/Examples/](https://github.com/hemanthpruthvi/PyAstroPol/tree/master/Examples) contains several examples files to demonstrate the applications of the package. They are provided in the form of `IPython` notebook and it is a good way to quick-start using the package. __They also function as the test cases__. 
+[PyAstroPol/Examples/](https://github.com/hemanthpruthvi/PyAstroPol/tree/master/Examples) contains several examples files to demonstrate the applications of the package. They are provided in the form of `IPython` notebook and it is a good way to quick-start using the package. **They also function as the test cases**. 
 
 ### Analysis on own
 
@@ -49,29 +45,32 @@ As previously mentioned, this is not a design software. Hence, one needs to know
 3. Detector   
 
 Following steps illustrate how to devise a simple optical system.  
-1. import the required package.
+1. import the required modules.
 ```python
-from PyAstroPol import * 
+import numpy as np
+from matplotlib import pyplot as plt
+from mpl_toolkits.mplot3d import Axes3D
+import PyAstroPol as pap
 ```  
 2. Create a source, and optionally create a source for display. For analysis one can define a source with a lot of rays (say 10000), and for display one can define a source with fewer rays (say 10).  
 ```python
-S_analysis = Source(10000, Clear=20)   # Source for analysis, with 10k rays and 20 mm size
-S_display = Source(10, Clear=20)       # Source for disply, with 10 rays and 20 mm size
+S_analysis = pap.Source(10000, Clear=20)        # Source for analysis, with 10k rays and 20 mm size
+S_display = pap.Source(10, Clear=20)            # Source for disply, with 10 rays and 20 mm size
 ```  
 3. Create a component such as surface, lens etc., and position it. 
 ```python
[…]
