### 2020-10-17 145281e
Modified documentation
