### 2020-10-17 63cfb19
Modified documentation
