### 2020-10-17 f5c4e3a
Modified documentation
