### 2020-10-17 4e893c7
Modified documentation
