+++ Docs/docs/conf.py
@@ -25,7 +25,7 @@ author = 'Hemanth Pruthvi'
 
 # The full version, including alpha/beta/rc tags
 release = '0.1.0'
-
+master_doc = 'index'
 
 # -- General configuration ---------------------------------------------------
 
