### 2020-10-17 ff7ddc4
Modified documentation
