+++ Docs/docs/conf.py
@@ -12,6 +12,8 @@
 #
 import os
 import sys
+sys.path.insert(0, os.path.abspath('../'))
+sys.path.insert(0, os.path.abspath('../../'))
 sys.path.insert(0, os.path.abspath('../../PyAstroPol/'))
 import sphinx_rtd_theme
 
