### 2020-10-17 50b129b
Modified documentation
