### 2020-10-17 5cf2796
Modified documentation
