+++ Docs/docs/conf.py
@@ -12,9 +12,9 @@
 #
 import os
 import sys
-sys.path.insert(0, os.path.abspath('./../'))
-sys.path.insert(0, os.path.abspath('./../../'))
-sys.path.insert(0, os.path.abspath('./../../PyAstroPol/'))
+sys.path.insert(0, os.path.abspath('../'))
+sys.path.insert(0, os.path.abspath('../../'))
+sys.path.insert(0, os.path.abspath('../../PyAstroPol/'))
 import sphinx_rtd_theme
 
 html_theme = "sphinx_rtd_theme"
