+++ README.md
@@ -91,7 +91,7 @@ print(MM)
 ## Directories
 
 [PyAstroPol/PyAstroPol/](https://github.com/hemanthpruthvi/PyAstroPol/tree/master/PyAstroPol)  
-It is the main directory containing all the source files. **For more information on the `Classes` check the documentation hosted at [ReadTheDocs](https://pyastropol.readthedocs.io/en/latest/)**.
+It is the main directory containing all the source files. **For more information on the `Classes` check the documentation hosted at [ReadTheDocs](https://pyastropol.readthedocs.io/index.html)**.
 
 [PyAstroPol/Materials/](https://github.com/hemanthpruthvi/PyAstroPol/tree/master/Materials)  
 It has the refractive index data for different materials in a formatted manner. These files are loaded by the code to look-up the refractive index information of the given material. Users can easily create such files using following steps.
