### 2020-10-17 c54ed5a
Modified documentation
