+++ README.md
@@ -10,6 +10,7 @@ The package imports following external libraries, all of which are ubiquitous. T
 numpy
 matplotlib
 ```
+**Documentation on the `Classes` is hosted at [ReadTheDocs](https://pyastropol.readthedocs.io/index.html)**.
 
 ## Getting Started
 
@@ -91,7 +92,7 @@ print(MM)
 ## Directories
 
 [PyAstroPol/PyAstroPol/](https://github.com/hemanthpruthvi/PyAstroPol/tree/master/PyAstroPol)  
-It is the main directory containing all the source files. **For more information on the `Classes` check the documentation hosted at [ReadTheDocs](https://pyastropol.readthedocs.io/index.html)**.
+It is the main directory containing all the source files.
 
 [PyAstroPol/Materials/](https://github.com/hemanthpruthvi/PyAstroPol/tree/master/Materials)  
 It has the refractive index data for different materials in a formatted manner. These files are loaded by the code to look-up the refractive index information of the given material. Users can easily create such files using following steps.
