### 2020-10-17 #2693 @hemanthpruthvi
@aquilesC  Thank you again for the review. Please find the updated contents.

The Paper
-----------

I have put a little more emphasis on the features and added some content as to how the code fits in the community. I have also re-written parts of the paper.

I believe that an explanation is needed regarding the writing choices. As I have mentioned in the updated paper, I do not know of an open-source software with the features of `PyAstroPol`. My counterparts in the field and I have followed the approach that is mentioned in the paper: make instrument specific models, use commercial software, or a combination of both. Hence, my first instinct was to compare with the commercial software that we have already been using, and the emphasis on the limitations is probably a reflection of that process.

The Code
-----------

I have added `requirements.txt` to the package. I have also elaborated on the installation instructions.

I have hosted the documentation at `ReadTheDocs` and provided the link in the `README`. I am grateful for your compliments regarding the code documentation. I am glad that you took notice of the efforts.

I have modified the examples to reflect the recommended usage of `import`. Now, it is `import PyAstroPol as pap`. I have some additional development plans which might require some restructuring. I will consider changes to the internal code at that time.

I have tried to adhere to lowerCamelCase for the most part, and all the methods follow that. Unfortunately, the attributes do deviate from this because I chose names/abbreviations which are popularly in use already. That is the case with the attributes such as `self.rs` or `self.Wavelength`. In case of `self.iRI`, the alternative was `self.iRefractiveIndex` which, I believe, would be needlessly longer.

Runtime warning
-------------------
I have now suppressed the runtime warning. The code uses array operations instead of conditional loops for the sake of performance. Hence, these warnings are inevitable, and I have taken special care to deal with such cases. It can be found in the form of occasional `nan_to_num` operations, and so the output results are proper.

The example named `11_KTTCoelostat` is provided so that the user can verify the accuracy claims. It compares the outputs of `PyAstroPol` and `Zemax OpticStudio`. Unfortunately, this is possible only if the user has the software license, and an additional package `PyZDDE (Python-Zemax interface)`.
