### 2020-10-19 #2693 @aquilesC
@hemanthpruthvi , good work! I've just checked the manuscript, and it sounds much better with the positive aspects of your software stated first, and also the improvements to your code and examples are definitely a good addition.

I have marked all the checkboxes, @pibion , I think the work is ready to be published, I believe my review concludes, but please feel free to ping me if I can do anything else for you.
