### 2020-10-22 #2693 @mwcraig
@pibion -- I'll get more comments in tonight.

@hemanthpruthvi -- neat looking package and thanks for the improvements so far!

1. My main concern at this point is installability of the package. Would you be amenable to adding a small setup.py file to your repository? I'm happy to help you get that written. I think many users will be hesitant to modify `PYTHONPATH` as it is no longer a standard practice and may have unintended side effects.
2. With the addition of a line or two of code in each of the Python files in `PyAstroPol` defining which things get import with `*` you could ensure that when users import your package they get only your modules, not additional ones like numpy, etc. I'm happy to help you do this too if you are amenable to doing it.
3. The addition of a `requirements` file is excellent! It looks like it has two `.txt` extensions, though, so you should probably rename it to `requirements.txt`.
4. There are a few folders in the repo that you could get rid of by adding them to `.gitignore` then deleting the folders and commit those changes. The ones I have in mind are any `.ipynb_checkpoints` and `__pycache__` folders.

The spirit of this is to help you make it as easy as possible for potential users to install and try out your package when the paper is published. The package itself is very interesting!
