### 2020-10-26 a41f532
Minor organisational mods
