### 2020-10-26 #2693 @hemanthpruthvi
@mwcraig Thank you very much for the review and the offer to help.

1. I very much agree with you, and with @aquilesC and @caldarolamartin  as well, regarding the installability.  I absolutely intended/intend to have a ``setup.py`` so that the users can install the package by simply using ``pip``, like all other packages. The reason it is not there at the moment is something to do with the *user data*.
The commercial software that is widely used for these purposes creates two directories when installed: *root directory* -- for the program files, and  *user directory* -- for the files to be accessed by the user. In my package, ``./Examples/``  and ``./Materials/`` come under the *user data*, some of which will also be accessed by the program in the runtime.
I think the simpler solution is to just keep one directory, but then it is not desirable to have it somewhere in ``site-packages``. Hence, I have resorted to present installation scheme. I was hoping to get some user input and features in, so that I can decide if I should take the trouble to have two directories.
I very much welcome if you have any ideas in that regard.

2. I have modified the file ``__init__.py``. Now, other packages like ``numpy`` will not be imported when ``PyAstroPol`` is imported.  accordingly.

3. Taken care.

4. Taken care.
