### 2020-10-26 #2693 @aquilesC
@hemanthpruthvi , what you mention does not have a unique answer, it will depend a bit on what you expect the users to achieve. There are two things you can do. One is, during installation create a folder in the $HOME directory (or create it the first time the package runs) and put the examples and materials in that folder, so the user can easily modify them.

You can also use a variable (environmental or explicit, something like: ``os.environ['pyastropol_folder']``) to point to the directory where the materials are stored.

You can also opt for installing your package in development mode (``pip install -e`` or ``python setup.py develop`` which will add to the PATH without the need to ask the user to explicitly mess with the Pythonpath, and you'll get the extra advantage of working *inside* virtual environments out of the box.

Although I think any of the suggestions above can be good additions from a user perspective, I don't think they are part of the review process itself.
