### 2020-10-27 3a87786
Additions to readme
