### 2020-10-27 #2693 @hemanthpruthvi
@aquilesC Thank you very much for all your suggestions! I will be considering them when revising the installation scheme.

@mwcraig I have appended README with a request to the users about the feedback regarding the installation sheme.
