### 2020-11-06 #2693 @mwcraig
> @mwcraig have @hemanthpruthvi's changes to the README satisfied the installation requirement, in your view?

Yes, given the special installation situation here.
