### 2020-11-06 #2693 @pibion
Hi @hemanthpruthvi  it looks like we're nearly ready to publish.  I'm reading over the paper and have a few comments.  I haven't finished reading the paper yet but wanted to get these to you quickly, I'm expecting to finish a bit later today!

## Questions about the paper.
- You say, "It can be easily distributed along with the polarization calibration routines."  It's not clear what "the polarization calibration routines" are - is this something every telescope has?
- You say, "The one simple goal is to calculate the Mueller matrix" - is this its only goal, or is this an example of the functions it makes available?
- I'm not entirely sure what you mean by "In the polarization analysis of the astronomical telescopes, various approaches have been adopted considering the complexity of the system."  Do you mean: Because astronomical telescopes are complex systems, many approaches to polarization analysis have been adopted?

## Suggestions for the text - feel free to ignore or incorporate at your discretion.
- The majority of the optical surfaces in the astronomical instruments induce a change in the state of polarization => The majority of the optical surfaces in the astronomical instruments change the state of polarization
- polarimetric instruments inevitably require the information of polarization properties => polarimetric instruments inevitably require information about the polarization properties
-  It can be determined through experiments => This can be determined through experiments
- The Python package PyAstroPol provides a means to analyze the polarization properties => The Python package PyAstroPol analyzes the polarization properties
-  Zemax – a commercial software is used =>  Zemax – a commercial software -- is used
