### 2020-11-06 #2693 @pibion
@hemanthpruthvi here are the rest of my comments on the paper.

## Questions about the paper
- It's not clear what you mean by "facilitate" when you say, "Off-axis components are facilitated, as they have a significant effect on polarization."  Do you mean calculate?  Or include?
- You say, " Spot diagram is possible at any instance" and I'm not sure what you mean by "instance."  Do you mean before and after each optical element?
- You say, "Only circular optics (apertures) can be devised at the moment" and I'm not sure that "devised" is the right word.  Do you mean "modeled?"

## Suggestions for the text
- PyAstroPol aims to provide better parts of the aforementioned approaches => PyAstroPol aims to combine the better parts of the aforementioned programs
- open-source tools developed using scientific programming language, with a range of applications in modelling the
astronomical optics => it is an open-source tool written in the Python language and offers a range of functions to model astronomical optics
- I would suggest using the latex "itemize" environment for your list of "Salient features" and "Important features"
- Calculation of the Mueller matrix of a given optical system is a simple and only end goal. Rest
are the by-products of such analysis. The system Mueller matrix is computed by coherently
adding the electric field vectors after propagation. It implies that the system is imaging type,
which is commonplace in astronomy. => PyAstroPol calculates the Mueller matrix of a given, imaging-type optical system.  All the by-products of this calculation are made available to users who wish to do further analysis.
-  Effect of multi-layered coatings, such as oxide layers and protective coatings, on the state
of polarization, is included => We include the effect of multi-layered coatings, such as oxide layers and protective coatings, on the state of polarization
-  Incidentally, birefringent components are not included yet. => Birefringent components are not included yet
- The visualization features are limited. - It is neither a design software nor interactive. => The visualization features are limited. - PyAstroPol is neither a design software nor interactive.
