+++ Docs/paper.md
@@ -23,32 +23,33 @@ bibliography: paper.bib
 
 # Statement of Need
 
-The instrumental polarization analysis is one of the key aspects of the optical system analysis of the astronomical telescopes and instruments. The majority of the optical surfaces in the astronomical instruments induce a change in the state of polarization of the incoming light, which could lead to inaccurate measurements of the state of polarization or _polarimetry_. Hence, polarimetric instruments inevitably require the information of polarization properties of the optical system or the _instrumental polarization_. It can be determined through experiments, analysis or a combination of both. As the telescopes become larger and instruments become more complex, the instrumental polarization analysis has become all the more crucial (e.g.,  @TMT_2015 ; @DKIST_2016).  
+The instrumental polarization analysis is one of the key aspects of the optical system analysis of the astronomical telescopes and instruments. The majority of the optical surfaces in the astronomical instruments induce a change in the state of polarization of the incoming light, which could lead to inaccurate measurements of the state of polarization or _polarimetry_. Hence, polarimetric instruments inevitably require the information about the polarization properties of the optical system or the _instrumental polarization_. This can be determined through experiments, analysis or a combination of both. As the telescopes become larger and instruments become more complex, the instrumental polarization analysis has become all the more crucial (e.g.,  @TMT_2015 ; @DKIST_2016).  
 
 # Summary
 
-The Python package `PyAstroPol` provides a means to analyze the polarization properties of a given optical system with relative ease and minimal dependencies. The one simple goal is to calculate the Mueller matrix (e.g., @Gil_2016) of the optical system.  
-It can be easily distributed along with the polarization calibration routines. 
+The Python package `PyAstroPol` provides a means to analyze the polarization properties of a given optical system with relative ease and minimal dependencies. The simple end goal is to calculate the Mueller matrix (e.g., @Gil_2016) of the optical system.  
+It can be easily distributed along with the polarimetric calibration routines.
 
-In the polarization analysis of the astronomical telescopes, various approaches have been adopted considering the complexity of the system. A significant part of the complexity is because the instrumental polarization is often time-dependent. For the solar telescopes with Coelostats, Mueller matrices were analytically derived as a function of time (e.g., @KTT_1985, @VTT_2005). For Thirty Meter Telescope (TMT), a combination of analytical and numerical methods is used (@TMT_2015). However, for Daniel K. Inouye Solar Telescope (DKIST), Zemax – a commercial software is used, along with the team's in-house tools (@DKIST_2016).
+In the polarization analysis of the astronomical telescopes, various approaches have been adopted depending on the complexity of the system. A significant part of the complexity is because the instrumental polarization is often time-dependent. For the solar telescopes with Coelostats, Mueller matrices were analytically derived as a function of time (e.g., @KTT_1985, @VTT_2005). For Thirty Meter Telescope (TMT), a combination of analytical and numerical methods is used (@TMT_2015). However, for Daniel K. Inouye Solar Telescope (DKIST), Zemax – a commercial software – is used, along with the team's in-house tools (@DKIST_2016).
 
-`PyAstroPol` aims to provide better parts of the aforementioned approaches: open-source tools developed using scientific programming language, with a range of applications in modelling the astronomical optics. There is no open-source software which has polarization ray propagation features as per my knowledge. Hence, Zemax OpticStudio&reg; modelling is used for comparison. The salient features and important limitations are listed below.
+`PyAstroPol` aims to combine the better parts of the aforementioned programs: it is an open-source tool written in the `Python` programming language and offers a range of functions to model the astronomical optics. There is no open-source software which has polarization ray propagation features as per my knowledge. Hence, Zemax OpticStudio&reg; modelling is used for comparison. The salient features and important limitations are listed below.
 
 Salient features :   
[…]
