### 2020-11-10 #2693 @hemanthpruthvi
@pibion  Thank you very much for your feedback. I have incorporated some of your suggestions for the text in the updated manuscript.

## Answers to questions (in the same order) :

* Replaced 'polarization calibration' with 'polarimetric calibration', which I think is more appropriate.
Polarimetric calibration routines use calibration data and modelling to correct the effects on the state of polarization caused by the telescope and the instrument.

* Edited the text 'The one simple' to 'The simple end'.
Mueller matrix computation is its only end goal i.e., only the features required to compute the Mueller matrix are developed. Since complete polarized ray propagation is required for that, a lot of other functions are also available as byproducts.

* Replaced 'considering' with 'depending on'.
There are various levels of complexities in the astronomical systems. For some simpler systems, instrumental polarization was derived analytically. For newer complex systems, significantly complex modelling including numerical methods is being used. The choice/approach as to how much modelling is to be used to compute the instrumental polarization is selected considering the complexity of the system.

* Replaced 'facilitated' with 'included'.

* Appended the text 'of the beam path' to 'at any instance'.
I meant to say spot diagram at any instance of the beam path, including but not limited to before and after each elements. I have modified the text accordingly.

* Replaced 'devised' with 'modelled'.
