### 2020-11-11 #2693 @pibion
@hemanthpruthvi okay, thanks for these updates!  I have only one question remaining:

I'm still not sure what you mean when you say "It can be easily distributed along with the polarimetric calibration routines."  Do you mean, "It can be easily distributed along with the polarimetric calibration routines for a particular telescope?"
