### 2020-11-13 #2693 @hemanthpruthvi
@pibion I mean that it can be easily distributed along with polarimetric calibration routines *in general*, provided that the pipeline is in/can use Python. The telescopes with widely different applications still use very similar optics, essentially, coated surfaces like mirrors and lenses. Users can make the telescope models using just `PyAstroPol`. Such models can be deployed as part of the polarimetric calibration routines of the pipeline, for when telescope polarization computations are required.
One of the provided examples `11_KTTCoelostat.ipynb` is a model for part of an existing telescope.
