### 2020-11-13 #2693 @pibion
Ah okay, that's clear.  What about, "It can be easily included alongside polarimetric calibration routines in any pipeline that has access to Python?"
