### 2020-11-14 #2693 @hemanthpruthvi
@pibion Although in recent times most pipelines are being developed in `Python`, they can be written using, popularly,  `IDL`, `IRAF` or `Java` etc. In such cases, these pipelines should be able to call `Python` routines.
