### 2020-11-16 #2693 @pibion
@hemanthpruthvi great, that's very clear.  Could you update the paper to clarify?  When you refer to "the calibration routines" it makes me wonder if you're referring to something that exists in this package.

Maybe something like, "Data pipelines can include this packages alongside e.g. polarimetric calibration routines."
