+++ README.md
@@ -1,6 +1,8 @@
 # PyAstroPol
 Instrumental Polarization Analysis of Astronomical Optics
 
+[![DOI](https://zenodo.org/badge/288481693.svg)](https://zenodo.org/badge/latestdoi/288481693)
+
 ## Overview
 The package has one simple goal : compute 4x4 Mueller matrix for the given optical system, and it is developed keeping astronomical optics in view.
 It uses geometric optics approach i.e., all the analysis uses strictly ray treatment. Users should keep in mind that this is NOT for optical design i.e., it is presumed that the user already knows the optical system that is to be analyzed.
@@ -11,6 +13,9 @@ numpy
 matplotlib
 ```
 **Documentation on the `Classes` is hosted at [ReadTheDocs](https://pyastropol.readthedocs.io/index.html)**.
+
+
+
 
 ## Getting Started
 
