### 2020-11-24 #2693 @arfon
@hemanthpruthvi - I made a few formatting fixes in this PR (mostly related to citations): https://github.com/hemanthpruthvi/PyAstroPol/pull/2

Once you've merged this, I will go ahead and publish.
